refactor: consolidate type imports and remove deprecated generic types
