+++ packages/lix-sdk/src/thread/create-thread-comment.ts
@@ -1,6 +1,6 @@
 import type { Lix } from "../lix/open-lix.js";
 import { nanoid } from "../database/nano-id.js";
-import type { NewState, State } from "../entity-views/generic-types.js";
+import type { NewState, State } from "../entity-views/types.js";
 import type { ThreadComment } from "./schema.js";
 
 export async function createThreadComment(
