+++ packages/lix-sdk/src/thread/create-thread.ts
@@ -1,7 +1,7 @@
 import type { Lix } from "../lix/open-lix.js";
 import type { Thread, ThreadComment } from "./schema.js";
 import { nanoid } from "../database/nano-id.js";
-import type { NewState } from "../entity-views/generic-types.js";
+import type { NewState } from "../entity-views/types.js";
 
 export async function createThread(args: {
 	lix: Lix;
